Modified directional rumble test
I modified the rumble test by playing the effects in shorter intervals
and repeating a complete circle multiple times.

It looks like the effect is gone now, I cannot really feel how it walks
from left to right.

Also, directional rumble in Wine joy.cpl doesn't work. But I did not
check if it is not implemented there or uses a different interface.