Pretend we are in Windows wireless mode
This commit removes the option for faking the device version number as
the HTML5 game controller APIs doesn't look at that (in contrast to
SDL which does) and instead pretends we are in Windows wireless mode.

That way, we still expose that the controller is connected in wireless
mode (so it shows up as "Xbox Wireless Controller" in softwares) but
we alleviate the several fixups already implemented in multiple API
layers.

Fixes: commit 43faf20e3bc08fdf1fc5575f19118da02c41339c
Fixes: commit 71adbbb5ac333130bb91a7 ("switched from 0x02EA to 0x02E0")
Fixes: commit 3f3860bc ("Add Support For Xbox Elite Series 2 Wireless")
Fixes: https://github.com/atar-axis/xpadneo/issues/113
Signed-off-by: Kai Krakow <kai@kaishome.de>