hid-xpadneo: Use work queue for battery events
Searching the Internet reveals that some people observe problems with
the controller due to battery reading. So let's make this asynchronous
and move it to a worker queue. This removes another spinlock from the
event hook.

Additionally, it is no longer parsed as part of the raw input event,
but instead within the HID event processor. The status itself will
be evaluated in a worker now since we need to hold a spin-lock on
xdata.

Signed-off-by: Kai Krakow <kai@kaishome.de>