hid-xpadneo: Convert mapping to using tables instead of code
This makes a much more convenient view of the mapping tables and uses
the same strategy as most kernel drivers do. The new tables allow for
adding more device variants more easily. With this commit we can
discard a vast amount of complex code.

We are now exposing the controller 100% compatible with Windows mode,
regardless of whether it is in Linux or Windows connection mode. Thus
we no longer need to jump through the hoops of trying to mix and
match the physical button and axis maps with the HID input layer
symbols. Wine games should be able to use hidraw for the controller
now (a lower latency connection mode as it skips SDL and several
other layers).

It also removes the descriptor dumps as these are in separate text
files now.

Maybe-related: https://github.com/atar-axis/xpadneo/issues/57
Maybe-affects: https://github.com/atar-axis/xpadneo/issues/180
Fixes: https://github.com/atar-axis/xpadneo/issues/162
Fixes: https://github.com/atar-axis/xpadneo/issues/177
Fixes: https://github.com/atar-axis/xpadneo/issues/182
Signed-off-by: Kai Krakow <kai@kaishome.de>