hid-xpadneo: Rework the directional rumble model
The current implementation only considered the upper half of the
directions circle. This design has some downsides:

  * It is not possible to use the trigger motors without using the
    main rumble motors

  * It is not possible to damp the main rumble compared to the
    trigger motors

The new design uses a linear approach instead. The forward gain of
the direction will ramp down the main rumble motors. The combined
forward and left gain will ramp up the left trigger motor in the lower
left quarter, and ramp up the right trigger motor in the upper left
quarter. For the right trigger motor, we apply the reverse strategy.

Signed-off-by: Kai Krakow <kai@kaishome.de>