hid-xpadneo: Use work queue for rumble effects
This makes the rumble playback asynchronous from the playback callback.

This reworks much of the code-design and replaces the previous patch
for sending rumble reports only when there's a rumble command to do.
With this change, we now get the following kernel message immediately
on disconnects:

    xpadneo 0005:045E:02FD.0054: failed to send FF report

It also renames the motor usage from left and right to weak and strong
because no one knows if the strong motor is left or right, it also
prevents confusion with the trigger motors which are clearly left or
right.

The general design is a valuable effort on improving code quality, as
it simplifies some complex functions.

Maybe-related: https://github.com/atar-axis/xpadneo/issues/171
Maybe-related: https://github.com/atar-axis/xpadneo/issues/122
Maybe-affects: https://github.com/atar-axis/xpadneo/issues/180
Signed-off-by: Kai Krakow <kai@kaishome.de>