Follow the kernel code style better
This adds indent rules applying the kernel code style recommendations
and fixes up most of the style issues in the code.

It also cleans up some logic to be simpler to read and declares more
functions as static.

The original coding style is defined in in the Linux source tree at
`/usr/src/linux/scripts/Lindent`.

Signed-off-by: Kai Krakow <kai@kaishome.de>