hid-xpadneo: Autodetect battery presence after connect
Now, we are seeing a unique ID. Additionally, the device won't report
a battery when such a report was never sent by the controller. We now
detect the report ID automatically instead of hard-coding it to 4.

The power supply device will now switch the type. As the device no
longer reports capacity when connected via USB, we remember the last
known capacity and report that instead.

Apparently, switching types dynamically seems now to be picked up by
at least KDE Plasma. The sysfs interface sees the change, tho. As a
work-around we pretend battery mode during initialization.

Signed-off-by: Kai Krakow <kai@kaishome.de>