hid_xpadneo: Add quirk modes for misbehaving controllers
This commit introduces a new parameter "quirks" to adjust the
protocol to work with non-standard controllers. To use this feature,
add a module parameter "quirks=MAC:number" where MAC is the Bluetooth
MAC of your controller (format 11:22:33:44:55:66) followed by ":" and
a sum of flags to enable.

This commit also features matching controllers by name to automatically
detect and set the quirks.

Disabling of trigger rumble is moved to the quirk handler without
writing the incident to dmesg. The effect of both functions is
basically the same.

Signed-off-by: Kai Krakow <kai@kaishome.de>