hid-xpadneo, events: Do not send Xbox logo event for turn-off
Currently, the controller sends a key-down event when trying to turn
the controller off. We should not do this, as it looks like the user
actually pressed a key.

This commit caches the key-down event until its release is detected.
If it is not, the controller turned off and no event was sent to
user space. If the user releases the button, we just replay the
cached event immediately followed by a key-up event.

Signed-off-by: Kai Krakow <kai@kaishome.de>