hid-xpadneo: Adhere to Linux Gamepad Specification
Linux Gamepad Specification defines the idle position of the joystick
to have values (0,0). Let's add a parameter for this and turn the
feature on by default.

This may negatively impact a few Wine games which access the joystick
directly. You may also need the reset calibration data of `jscal` by
opening `/var/lib/joystick/joystick.state` and removing the line
prefixed with `jscal -s` while the controller is turned off. If the
file does not exist or is empty, there's nothing to fix.

Fixes: https://github.com/atar-axis/xpadneo/issues/57

This reverts commit 232a7321e5d8b04836448a75af4f9f01e923291f.

Signed-off-by: Kai Krakow <kai@kaishome.de>