udev: Expose all xpadneo input devices as user-readable
This allows hidraw applications to read the device directly without
going through the input layer. This may be useful to certain FPS games
that allow a low-latency input mode by using this. Our hidraw byte
stream is compatible with the original Windows Xbox One S HID mode, so
we don't need to hide it.

Signed-off-by: Kai Krakow <kai@kaishome.de>