hid-xpadneo, battery: Rework detection, parsing and reporting
The battery is now registered lazily so upower no longer has a chance
to read an unknown or zero-level battery state. By coincidence, we can
now also detect the controller shutting down.

The capacity levels have been adjusted to no longer report critical
because even in the lowest charge state the batteries still last
quite a time.

Additionally, we now report if you're using the Play 'n Charge Kit by
showing different battery names in desktop notifications.

Keep in mind that the controller won't tell us the capacity level
while USB is connected, so we cannot report the charging progress.

Also keep in mind that the upower charge history may show 0 but this
is okay as upower is not supposed to use the percentage value, so
you shouldn't either.

This commit hasn't been tested with the Play'n Charge Kit.

Fixes: https://github.com/atar-axis/xpadneo/issues/212
Signed-off-by: Kai Krakow <kai@kaishome.de>