hid-xpadneo, quirks: We need to carry a quirk for Linux button mappings
There are controllers which are not 100% compatible with Xbox One S
because they are not switching to Linux mode when connected to Linux.
We can detect such devices at HID parse time and set a flag when we
detect Linux mode.

With this commit, we no longer shift the button bits when we detect a
controller in native Windows mode.

Fixes: https://github.com/atar-axis/xpadneo/issues/228
Signed-off-by: Kai Krakow <kai@kaishome.de>