xpadneo, deadzones: Use smaller dead zone and fuzz for precision
These values have been determined experimentally on a medium worn
controller. This also removes dead zones completely for the triggers as
the controller already implements that in the firmware.

Fuzz is shrunk by a factor of 8 because the sticks are very precise
already and we don't want to filter out this precision. We can set the
dead zone 25% smaller compared to Linux defaults.

See-also: https://github.com/atar-axis/xpadneo/issues/231
Signed-off-by: Kai Krakow <kai@kaishome.de>