xpadneo, deadzones: Implement a high-precision mode without dead zones
This adds a module parameter `disable_deadzones` to remove dead zone
handling from upper layers. This is only useful if you're a playing
games in Wine or Proton as otherwise SDL introduces a dead zone handler
which seriously hurts precision compared to native Windows usage of the
controller.

Do not use this if you intend to play native games that make use of the
`joydev` API in Linux as the removal of the dead zones introduces
axis shift and phantom movement. In Windows, games are supposed to
handle dead zones by themselves while in Linux the drivers to this.

Fixes: https://github.com/atar-axis/xpadneo/issues/231
Signed-off-by: Kai Krakow <kai@kaishome.de>