hid-xpadneo, rumble: Do not lose rumble strength while throttled
While the rumble thread is throttled, we may actually miss the peak
rumble strength resulting in an experience of rumbles with varying
strength when these are repeatedly sent by a game.

To work around this, we now accumulate the maximum strength set during
the throttle interval, then reset the accumulator when a rumble has
been sent to the controller.

Maybe-fixes: https://github.com/atar-axis/xpadneo/issues/290
Signed-off-by: Kai Krakow <kai@kaishome.de>