hid-xpadneo: Allow the driver to operate in Xpad360 emulation mode
This mode disables most advanced features and exposes the device as an
original USB-connected Xbox 360 controller. Use module parameter
`xpad_emulation=1`.

This may be useful for applications like Retroarch or other programs
that need the Guide button exposed as a joystick button. Essentially,
it enables such applications to use the button as a macro modifier
button, and the button will no longer be sent as a consumer control
key (but still announced as being available).

In this mode, joydev inserts the button before the left and right
thumb buttons which may mess up some mappings, especially with joydev.
Since we are pretending to be a USB Xbox360 controller, most programs
will take care of it.

For hidraw programs, we also scale the trigger range down from 0..1023
to 0..255 which makes the trigger slightly less precise.

This parameter cannot be changed at runtime, you must reload the module
with the parameter changed to apply the emulation correctly.

Fixes: https://github.com/atar-axis/xpadneo/issues/237
Affects: https://github.com/atar-axis/xpadneo/issues/273
Fixes: https://github.com/atar-axis/xpadneo/issues/280
Signed-off-by: Kai Krakow <kai@kaishome.de>