dkms: remove REMAKE_INITRD
By selecting this option, dkms attempts to add the module into the user
init ramdisk. While it seems like a cool idea it is not something the
module should decide by the user or distribution on their behalf.

I'm aiming to purge the option in upstream dkms, since it's fairly
convoluted and error-prone code. So let's remove it from this module.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>