xpadneo: Add paddles support
We still silence the paddles when one of the profiles is selected, the
default profile setup maps the paddles to A, B, X, and Y.

The paddles are special because they are passed as one key by the HID
report. We need to manually remap those values to individual bits.

Maybe-affects: https://github.com/atar-axis/xpadneo/issues/344
Closes: https://github.com/atar-axis/xpadneo/issues/332
Signed-off-by: Kai Krakow <kai@kaishome.de>