dkms: Suggest trusting the git directory if version detection failed
Git 2.35.2 added a trust model to prevent user scripts and hooks of git
repositories to be executed with elevated privileges by accident. This
prevents the installer from getting the repository version when the
repository was checked out by a non-root user.

This commit adds an error message to suggest trusting the repository
and also links to the bug report for further details.

Fixes: https://github.com/atar-axis/xpadneo/issues/346
Signed-off-by: Kai Krakow <kai@kaishome.de>