xpadneo: Add support for GuliKit KingKong2 PRO controllers
This adds a proper quirk handler for the GuliKit KingKong family of
controllers. These controllers use a reversed mapping of the motor
masking bits. This was tested with the KingKong2 PRO model but it
should generally work for the whole family of these controllers.

This adds a new quirk flag 128 to tell xpadneo about reversed motor
programming masks. GuliKit OUIs are automatically adjusted to use this
flag.

Since the GuliKit controllers allow swapping A/B and X/Y by holding
the settings button and pressing the plus button, xpadneo will default
to Xbox mappings for this controller, Nintendo mappings can be set as
default via a quirks flag.

Co-authored-by: Termy <saldorin@web.de>
Fixes: https://github.com/atar-axis/xpadneo/issues/391
Signed-off-by: Kai Krakow <kai@kaishome.de>