xpadneo, init: Do not report paddles unconditionally
SDL2 author @slouken pointed out that the bitmap cannot be properly
tested for the existence of the XBE2 paddles. This commit fixes that by
first testing if the paddles exist in the bitmap, and only then enable
the remaining three paddles. The underlying problem is that the HID
descriptor does not have a properly defined bitmap for the paddles but
a single 4-bit value instead.

See-also: https://github.com/atar-axis/xpadneo/issues/428
Signed-off-by: Kai Krakow <kai@kaishome.de>