xpadneo, init: Detect HW profile support on current firmwares
This makes it difficult to handle the existence of the Share button,
because all models ship it in the HID bitmap, and we can detect the
presence of HW profiles only late. Take note of that so it doesn't get
missed later when we remap that button because only XBE2 will allow us
to actually unmap it early from the bitmap (because it has a different
PID).

Signed-off-by: Kai Krakow <kai@kaishome.de>