core, quirks: Add GuliKit KK3 MAX quirks
This model has the masking bits for trigger and main motors swapped,
so essentially it doesn't rumble at all. This commit fixes it by adding
a new quirk flag. This quirk is very similar to the quirks needed by
the previous model series.

However, further testing showed that it doesn't properly obey the
motor masking bits it supports and won't stop the motors. So we need
to disable motor masking altogether which makes the swapped mask flag
quite useless. We keep that supported, tho, for future models.

Fixes: https://github.com/atar-axis/xpadneo/issues/490
Signed-off-by: Kai Krakow <kai@kaishome.de>