xpadneo: Add documentation about testing pull requests
Forks used for pull requests should not have tags. This is a problem
for our installer. To properly work with forks, these should be added
as a git remote instead. To still get the benefits of isolated
directories, git worktrees can be used instead.

Place a hint in the installer if this problem may occur, and also add
some documentation on how to use these features of git.

Signed-off-by: Kai Krakow <kai@kaishome.de>