Update hid-xpadneo.c
After the latest firmware update, rumble no longer worked on my Xbox Wireless Controller. It seems like the OUI has changed and xpadneo was incorrectly applying the GameSir Nova quirks. The following patch adds the OUI for the controller with no quirks. This was enough to bring rumble back to the controller.