xpadneo, udev: Trigger loading of `uhid` before `bluetooth`
The system should trigger loading of the `uhid` module when `bluetooth`
is loaded to properly support BLE devices which provide the HID bus
from user-space through bluez.

This may still leave a gap for a race-condition if both `bluetooth` and
the BLE device appear at the same time. `uhid` should be present before
any device even connects from user-space, otherwise `hid-xpadneo` may
not see the HID bus early enough.

I'm unsure if we also should provide `weakdep` to trigger inclusion of
`uhid` into initrd - but since Xbox controllers are not critical input
devices during early boot, this is probably not our business. OTOH,
`softdep` may already trigger inclusion on its own (untested).

Hint: This feels wrong because xpadneo is the HID consumer, not the HID
provider. We should not need to do that. But it prevents bug reports
from users in the future.

Tested-by: flodavid <fl.david.53@gmail.com>
Link: https://github.com/atar-axis/xpadneo/pull/563#issuecomment-3423054386
Signed-off-by: Kai Krakow <kai@kaishome.de>