xpadneo, dkms: Hooks are now effectively a no-op
DKMS hooks are now effectively a no-op and only remove a historical
artifact for old kernels that we handled outside of package management.
The file should actually no longer exist on any system.

Thus, we leave this in-place until one of the next releases, instead
adding a comment to explain the idea.

Fixes: https://github.com/atar-axis/xpadneo/issues/521
Signed-off-by: Kai Krakow <kai@kaishome.de>